requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.2
lark-oapi==1.4.5
qdrant-client==1.12.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.1.0
python-dotenv==1.0.1
tiktoken==0.8.0
numpy==1.26.4
rank-bm25==0.2.2
# Embedding & LLM
sentence-transformers==3.0.1
openai==1.51.0
dashscope==1.20.11
# 工具
loguru==0.7.2
tenacity==9.0.0